Refactored MoreInfoModal into Modular Components
- Extracted the modal code from MoreInfoModal.js and AccountHomePage into separate components under  components/moreInfoModal:

    1) ModalHeader.js - Handles the hero section (background image, gradient overlay,
      close button, title, buttons, and info rows with summary and extra details).
    2)  EpisodesList.js - Renders the list of episodes in a horizontal scroll format.
    3)  TrailersSection.js - Displays the "Trailers & More" section as a 3-column grid.
    4) AboutSection.js - Shows additional details such as Director, Cast, Genres, and
      maturity/disclaimer info.
    5) ProgramDetailsModal.js - Main modal component that assembles the above components.
   functionality and design remain the same, but the code is now more modular...
